major changes to gameMap class

+++ src/Point.h
@@ -1,6 +1,5 @@
 #pragma once
 
-
 struct Point {
-	short int x, y;
+	int x, y;
 };
